+++ civigiftaid.php
@@ -15,8 +15,13 @@ function civigiftaid_civicrm_config(&$config) {
   if (isset(Civi::$statics[__FUNCTION__])) { return; }
   Civi::$statics[__FUNCTION__] = 1;
 
+  // Symfony hook priorities - see https://docs.civicrm.org/dev/en/latest/hooks/usage/symfony/#priorities
   // Add listeners for CiviCRM hooks that might need altering by other scripts
   Civi::dispatcher()->addListener('hook_civicrm_post', 'CRM_Civigiftaid_SetContributionGiftAidEligibility::run');
+
+  // Run before giftaidonline (-100)
+  \Civi::dispatcher()->addListener('hook_civicrm_navigationMenu', 'civigiftaid_symfony_civicrm_navigationMenu', 0);
+
 }
 
 /**
@@ -98,11 +103,12 @@ function civigiftaid_civicrm_alterSettingsFolders(&$metaDataFolders = NULL) {
 /**
  * Add navigation for GiftAid under "Administer/CiviContribute" menu
  *
- * @param array $menu
+ * @param \Civi\Core\Event\GenericHookEvent $event
+ * @param string $hookName
  *
  * @throws \CiviCRM_API3_Exception
  */
-function civigiftaid_civicrm_navigationMenu(&$menu) {
+function civigiftaid_symfony_civicrm_navigationMenu($event, $hookName) {
   // Get optionvalue ID for basic rate tax setting
   $result = civicrm_api3('OptionValue', 'getsingle', ['name' => 'basic_rate_tax']);
   if ($result['id']) {
@@ -118,7 +124,7 @@ function civigiftaid_civicrm_navigationMenu(&$menu) {
     'operator'   => NULL,
     'separator'  => 1,
   ];
-  _civigiftaid_civix_insert_navigation_menu($menu, 'Administer/CiviContribute', $item[0]);
+  _civigiftaid_civix_insert_navigation_menu($event->params, 'Administer/CiviContribute', $item[0]);
 
   $item[] = [
     'label' => E::ts('GiftAid Basic Rate Tax'),
@@ -128,7 +134,7 @@ function civigiftaid_civicrm_navigationMenu(&$menu) {
     'operator'   => NULL,
     'separator'  => NULL,
   ];
-  _civigiftaid_civix_insert_navigation_menu($menu, 'Administer/CiviContribute/admin_giftaid', $item[1]);
+  _civigiftaid_civix_insert_navigation_menu($event->params, 'Administer/CiviContribute/admin_giftaid', $item[1]);
 
   $item[] = [
     'label'      => E::ts('Settings'),
@@ -138,9 +144,9 @@ function civigiftaid_civicrm_navigationMenu(&$menu) {
     'operator'   => NULL,
     'separator'  => NULL,
   ];
-  _civigiftaid_civix_insert_navigation_menu($menu, 'Administer/CiviContribute/admin_giftaid', $item[2]);
+  _civigiftaid_civix_insert_navigation_menu($event->params, 'Administer/CiviContribute/admin_giftaid', $item[2]);
 
-  _civigiftaid_civix_navigationMenu($menu);
+  _civigiftaid_civix_navigationMenu($event->params);
 }
 
 /**
@@ -238,23 +244,53 @@ function civigiftaid_civicrm_postProcess($formName, &$form) {
  * @throws \CiviCRM_API3_Exception
  */
 function civigiftaid_civicrm_post($op, $objectName, $objectId, &$objectRef) {
-  if ($objectName !== 'Contribution') {
-    return;
-  }
+  switch ($objectName) {
+    case 'Batch':
+      if ($op !== 'delete') {
+        return;
+      }
+      // We have the batch ID. Check if it's a giftaid batch and delete related stuff
+      try {
+        $optionGroupID = civicrm_api3('OptionGroup', 'getvalue', [
+          'name' => 'giftaid_batch_name',
+          'return' => 'id'
+        ]);
+        $batchOptionValue = civicrm_api3('OptionValue', 'getsingle', [
+          'option_group_id' => $optionGroupID,
+          'value' => $objectId,
+        ]);
+        // Clear batch_name from contributions
+        $updateSql = "UPDATE civicrm_value_gift_aid_submission SET batch_name = NULL WHERE batch_name=%1";
+        $updateSqlParams = [
+          1 => [$batchOptionValue['name'], 'String']
+        ];
+        CRM_Core_DAO::executeQuery($updateSql, $updateSqlParams);
+        // Finally we delete the giftaid batch name option value.
+        civicrm_api3('OptionValue', 'delete', [
+          'id' => $batchOptionValue['id'],
+        ]);
+      } catch (Exception $e) {
+        \Civi::log()
+          ->error('Deleting Gift Aid Batch failed: ' . $e->getMessage());
+      }
+      break;
 
-  if ($op == 'edit' || $op == 'create') {
-    $callbackParams = [
-      'entity' => $objectName,
-      'op' => $op,
-      'id' => $objectId,
-      'details' => $objectRef,
-    ];
-    if (CRM_Core_Transaction::isActive()) {
-      CRM_Core_Transaction::addCallback(CRM_Core_Transaction::PHASE_POST_COMMIT, 'civigiftaid_callback_civicrm_post_contribution', [$callbackParams]);
-    }
-    else {
-      civigiftaid_callback_civicrm_post_contribution($callbackParams);
-    }
+    case 'Contribution':
+      if ($op == 'edit' || $op == 'create') {
+        $callbackParams = [
+          'entity' => $objectName,
+          'op' => $op,
+          'id' => $objectId,
+          'details' => $objectRef,
+        ];
+        if (CRM_Core_Transaction::isActive()) {
+          CRM_Core_Transaction::addCallback(CRM_Core_Transaction::PHASE_POST_COMMIT, 'civigiftaid_callback_civicrm_post_contribution', [$callbackParams]);
+        }
+        else {
+          civigiftaid_callback_civicrm_post_contribution($callbackParams);
+        }
+      }
+      break;
   }
 }
 
@@ -379,5 +415,31 @@ function civigiftaid_civicrm_validateForm($formName, &$fields, &$files, &$form, 
 
   if (!empty($errors)) {
     return $errors;
+  }
+}
+
+/**
+ * This hook is called to alter Custom field value before its displayed.
+ *
+ * @param string $displayValue
+ * @param mixed $value
+ * @param int $entityId
+ * @param array $fieldInfo
+ */
+function civigiftaid_civicrm_alterCustomFieldDisplayValue(&$displayValue, $value, $entityId, $fieldInfo) {
+  // Gift Aid batch name is stored as "name" but we want to display "label".
+  if ($fieldInfo['name'] === 'Batch_Name' && $fieldInfo['column_name'] === 'batch_name' && !empty($value)) {
+    try {
+      $optionGroupID = civicrm_api3('OptionGroup', 'getvalue', ['name' => 'giftaid_batch_name', 'return' => 'id']);
+      $displayValue = civicrm_api3('OptionValue', 'getvalue', [
+        'option_group_id' => $optionGroupID,
+        'name' => $value,
+        'return' => 'label',
+      ]);
+    }
+    catch (Exception $e) {
+      // Do nothing, we'll use the existing displayValue
+      // This will fail for older batches which stored the label instead of the name for the batch_name field.
+    }
   }
 }